# server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mingle API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/reactions/{reactionId}/defaultConfig": {
            "get": {
                "tags": [
                    "Action / Reaction"
                ],
                "summary": "Get default config for a reaction",
                "operationId": "959dbf3dc0060b7d3c65aa8282d5022f",
                "parameters": [
                    {
                        "name": "actionId",
                        "in": "path",
                        "description": "ID of reaction that needs to be fetched",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                },
                                "example": "{}"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request ID need to be numeric."
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
...
    },
...
}